Other message

+++ src/events/dubious.ts
@@ -9,8 +9,10 @@ class DubiousFilter {
     private async onMessage([message]: ArgsOf<"messageCreate">, client: Client) {
         if (message.content.match(/jonathandata1/)) {
             console.log(`deleting message from ${message.author.username} in channel ${message.channel}`)
-            await message.delete()
-            await message.author.send(`Yo, had to delete that message from ${message.channel}`)
+            await message.author.send(`Yo, had to delete that message from ${message.channel} for policy reasons`)
+            if (message.deletable) {
+                await message.delete()
+            }
         }
     }
 }
